Doing a bajillion of things simultaneously. Here and there. Yet more and more unused stuff dropped. We're now under 1GB of RAM. Long way to go.
Those are the kind of things to be done the first day, not in the latest release.
Ignore the warnings in OpenBRF. I'm still moving assets.

This is a sort of stable backup in case than that the so called (and much needed) in-deep cleanup messes up.

I'm also planning to start the closed beta ASAP.

--swy